Restructure CLI
