--- crates/criticalup-cli/src/commands/mod.rs
@@ -1,14 +1,0 @@
-// SPDX-FileCopyrightText: The Ferrocene Developers
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-pub(crate) mod archive;
-pub(crate) mod auth;
-pub(crate) mod auth_remove;
-pub(crate) mod auth_set;
-pub(crate) mod clean;
-pub(crate) mod doc;
-pub(crate) mod install;
-pub(crate) mod remove;
-pub(crate) mod run;
-pub(crate) mod verify;
-pub(crate) mod which;
